월 상환가능금액 Long으로 통일

+++ scoula/backend/src/main/java/org/scoula/member/dto/UpdateMaxPaymentDTO.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UpdateMaxPaymentDTO {
-    private Integer maxMonthlyPayment;
+    private Long maxMonthlyPayment;
 }
